src/main/java/com/github/wenslo/springbootdemo/reposiroty/UserRepository.java
package com.github.wenslo.springbootdemo.reposiroty;

import com.github.wenslo.springbootdemo.model.User;
import org.springframework.data.jpa.repository.JpaRepository;
import org.springframework.data.querydsl.QuerydslPredicateExecutor;
import org.springframework.stereotype.Repository;

/**
 * @author wenhailin
 * @version 0.0.1
 * @createTime 2018年11月24日 下午3:59
 * @description
 */
@Repository
public interface UserRepository extends JpaRepository<User, Long>, QuerydslPredicateExecutor<User> {
    public User findByUsername(String username);

}
